Move utils to lib.utils submodule
There are two utils submodules right now,
so lets merge them together.
