+++ znoyder/tests/test_utils.py
@@ -18,10 +18,10 @@
 
 from unittest import TestCase
 
-from znoyder.utils import drop_nones_from_dict
-from znoyder.utils import match
-from znoyder.utils import merge_dicts
-from znoyder.utils import sort_dict_by_keys
+from znoyder.lib.utils import drop_nones_from_dict
+from znoyder.lib.utils import match
+from znoyder.lib.utils import merge_dicts
+from znoyder.lib.utils import sort_dict_by_keys
 
 
 class TestDropper(TestCase):
